olive/passes/onnx/ryzen_ai/examples/onnx/hugging_face_timm_quantization/requirements_gpu.txt
numpy
onnx>=1.16.0
onnxruntime-gpu
onnxruntime_extensions
Pillow
protobuf
rich
timm
